worker/requirements.txt
celery==5.5.0
redis==6.1.0
pillow==10.4.0
pydantic==2.10.5
numpy==1.24.0